pyproject.toml
[build-system]
requires = ["setuptools>=70.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hy_embodied"
version = "0.5.0"
description = "HY-Embodied-0.5-X: an embodied multimodal foundation model for real-world robotics."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "Tencent Robotics X Lab" },
    { name = "Tencent Hunyuan Team" },
]
keywords = [
    "embodied-ai",
    "multimodal",
    "vision-language",
    "robotics",
    "foundation-model",
    "hunyuan",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
# Runtime deps are intentionally kept lean here — full pinned lists live in
# requirements.txt and setup_env.sh.  We only declare what is strictly
# required to import the library and run basic inference.
dependencies = [
    "torch>=2.8",
    "transformers>=4.57.0",
    "accelerate>=1.0",
    "safetensors>=0.5",
    "Pillow",
    "PyYAML",
    "numpy",
]

[project.optional-dependencies]
train = [
    "deepspeed>=0.18",
    "liger-kernel>=0.6",
    "tensorboard>=2.20",
]
serve = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]
dev = [
    "pytest>=8",
    "ruff>=0.6",
]

[project.scripts]
hy-embodied-train = "hy_embodied.cli.train:main"
hy-embodied-infer = "hy_embodied.cli.infer:main"
hy-embodied-server = "hy_embodied.cli.server:main"

[project.urls]
Homepage = "https://github.com/Tencent/HY-Embodied-0.5-X"
Repository = "https://github.com/Tencent/HY-Embodied-0.5-X"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["hy_embodied*"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
